修复了数据库导入的问题。同时增加了一个标示目录的TextView。

+++ src/org/bitren/app/control/UtilControl.java
@@ -43,7 +43,7 @@ public class UtilControl {
 				ContactEntity contact = new ContactEntity();
 				JSONObject jsonObject = jsonArray.getJSONObject(i).getJSONObject("BitContact");
 				
-				contact.setSid(jsonObject.getLong("id"));
+				contact.setSid(jsonObject.getInt("id"));
 				contact.setDepartment(jsonObject.getString("department"));
 				contact.setLocation(jsonObject.getString("location"));
 				contact.setPhone_number(jsonObject.getString("phone_number"));
